refactor(workers): split cloud placement modules
